requirements.txt
Django==5.0.6
python-dotenv==1.0.1
psycopg2==2.9.9
pillow==10.3.0
colorama==0.4.6
drf-yasg==1.21.7
setuptools==70.0.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1